+++ src/hosts/vscode/terminal/VscodeTerminalRegistry.ts
