+++ src/hosts/vscode/terminal/ansiUtils.ts
