+++ src/integrations/terminal/StandaloneTerminal.ts
