+++ src/integrations/terminal/StandaloneTerminalManager.ts
