+++ src/integrations/terminal/StandaloneTerminalRegistry.ts
