+++ src/integrations/terminal/index.ts
