+++ src/integrations/terminal/types.ts
