+++ src/test/host-provider-test-utils.ts
@@ -7,7 +7,7 @@ import {
 } from "@/hosts/host-provider"
 import { HostBridgeClientProvider } from "@/hosts/host-provider-types"
 import { vscodeHostBridgeClient } from "@/hosts/vscode/hostbridge/client/host-grpc-client"
-import { ITerminalManager } from "@/shared/terminal/types"
+import { ITerminalManager } from "@/integrations/terminal/types"
 
 /**
  * Initializes the HostProvider with test defaults.
